Increase memory requests for Container Registry
The default is 32Mi and our service runs higher than that number if it
is not taking any traffic.  We are seeing some amount of Evicted Pods,
all due to memory pressure.  Use this as a method of impacting the
requests on a server to hopefully scale up the amount of nodes we need
for a cluster to run the desired amount of Pods.

The choice of running around 500MB as it is half of the current running
average across all Pods.  I'm still learning what an appropriate value
for this would be.  But based on this number, and based on the current
number of running Pods, 30 Pods x 500MB = 15GB requested memory.  Our
nodes have 7.5GB of RAM each, and with requests being asked for by other
Pods, this should trigger the Kubernetes Cluster to scale.  With more
nodes to run, we should see less memory pressure on our nodes leading to
less evicted Pods

This will change the memory requests amount of 30 Pods from 960Mi to 14280Mi