Limit the Container Registry memory usage
* This sets a very conservative amount of memory for which the Container
Registry is allowed to utilize
* For GitLab.com we appear to hover around the 1GB mark after a Pod has
been around for awhile.  Our standard deviation can easily breach 7-13GB
in size.
* Setting this to 6GB as we do have an EmptyDir mount of type Memory
  * This is pointed to our configuration file which contains only 12KB
  of data:
```
Filesystem      Size  Used Avail Use% Mounted on
tmpfs            15G   12K   15G   1% /etc/docker/registry
```
* Since this limit takes this EmptyDir usage into consideration but we
do not use this scratch space for the Container Registry, I believe it
is safe to assume that majority of the memory usage is consumed by the
running process and not the data in this scratch space
* Setting this limit will trigger Kubernetes to restart the Pod, which
should eliminate occurrences of Evicted Pods that we occasionally see in
the production environment