Revert/correct the instructions for adding the registry cert secret
As discovered in https://gitlab.com/gitlab-org/gitlab/issues/199896, the certificate in question is the rootcertbundle which is used to validate the tokens (JWT) generated by the rails application, and needs to match the 'internal_key' key that Rails is using to sign those tokens.  It is not the public certificate used for HTTPS access to the registry.  Additionally the registry only requires the cert, not any private key.