Fix for populating IP address
* The gcloud CLI has started to issue a warning when running the gcloud
cli with `filter` as an option

```
% gcloud --project gitlab-production compute addresses list --filter name=registry-gke-gprd --format value\(address\)
10.216.8.3
WARNING: --filter : operator evaluation is changing for consistency across Google APIs.  name=registry-gke-gprd currently does not match but will match in the near future.  Run `gcloud topic filters` for details.
```

This warning interferes with our ability to gather that data.  What's
odd is that this warning indicates it does not match today, but will in
the future, desire still returning the desired result, proving that the
filter is working as desired.

Because the error shows up in our output, helmfile attempts to process
this and fails.

This commit uses the global `verbosity` flag and sets it to `error`.
We'll use this setting until we notice a behavior change.