Fixes initial install of GitLab
* Without this, the namespace or the deployment may not exist, causing
kubernetes to exit prematurely which causes helmfile to stop
* With this, we ignore the failure from the `kubectl` side, which allows
the helm chart to complete populating the variable properly which in
turn allows a first time install of GitLab to complete
* note that this sets the initial value to that defined in the chart,
whether that be the version defined for a release, example 13.1.0, or
master when using a non tagged version of the chart.  In the latter
case, this will default to tag `latest`.