Add a single catchall sidekiq pod with rbtrace enabled
To debug the memory behavior seen in https://gitlab.com/gitlab-com/gl-infra/production/-/issues/2778 it would be really handy to have rbtrace enabled.  We only need a single such pod, so to minimize the potential for disruption, this creates a new class of sidekiq pod with ENABLE_RBTRACE set to 1, but only a single replica (min and max).  Once this is deployed, we can find that single pod and debug it as necessary.