Move ingress for kas to `kas.pre.gitlab.net` in `pre`
This is part of https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/11581

This involves setting up a new helm release called `gitlab-extras` in order to
add some GKE specific custom resources in order to make this happen. KAS will now
be

* Configured on subdomain `kas.pre.gitlab.com`
* Using a Google Managed Certificate as I didn't feel there was any need to keep
managing a certificate for this manually
* Using a GKE Ingress, which is implemented by a GCP HTTPS Loadbalancer. This means
it will be completely isolated from our ingress-nginx used for `gitlab-webservice`,
and no haproxy in front of it.
* Using [Container Native Load Balancing](https://cloud.google.com/kubernetes-engine/docs/concepts/container-native-load-balancing)
which allows us to have the load balancer talk directly to the pods, bypassing
kube-proxy and other un-needed network hops/complexity for improved performance.
* As KAS is a websocket/gRPC application, we will need a custom HTTP healthcheck
configured as the healthcheck endpoints do not live on the same port as the rest
of the traffic (HTTPS Load Balancers don't support websocket/grpc native health
checks).