Tweak the backendconfig configuration for `kas`
Fixes: https://gitlab.com/gitlab-org/gitlab/-/issues/284074

According to Google docs we currently timeout all websocket connections
at 30 seconds by default. As `kas` is is designed for longer lived
connections, we wish to up the timeout to reduce the number of
connection errors we are facing. As `kas` itself reconnects
every 30 minutes, we wish to set the timeout to 30 minutes plus
a little bit longer.

We also remove the specific checkInterval from this configuration
as 60s is way too long, and wish to go back to the default of 5s