Move logic around fetching chef role outside `helmfile`
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/699

Currently we cannot upgrade `helmfile`, as newer versions have changes
to the `exec` function which can accidentally leak the secret ops api
token in CI jobs (and other places). @skarbek lodged upstream feature
request at

https://github.com/roboll/helmfile/issues/1397

in order to make helmfile have the ability to not leak this secret
value. However, as that RFE hasn't been implemented yet, this change
instead moves the logic for fetching a chef role outside of helmfile
so that we won't accidentally leak the secret if something goes wrong.

This has the follow on effect of allowing us to unblock us from upgrading
to newer versions of helmfile as needed.