Explicitly set the `gitlab-extras` namespace for canary
Part of https://gitlab.com/gitlab-org/gitlab/-/issues/249596

While you would think that we wouldn't need to specify this as we
don't install the release in cny, what is happening is instead is that
the cny apply job is seeing the release is installed in the `gitlab`
namespace (that's the default), realises it needs to not be installed
(as we disable it in cny configuration), and thus removes it.

So by setting the namespace for canary to something different, canary
will check it's not installed in the canary namespace (which is correct),
instead of the production namespace.