Enable redis configuration for `kas` in pre and staging
Part of https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/12519

This enables the redis configuration of `kas`, which is needed to enable
per IP rate-limiting in the kas service. The data stored (and its size)
is in the linked issue above. We will be using the `redis-cache` cluster
to store this data as it doesn't need to be persisted.