Change consul ns record to be fully qualified in production.
As part of https://gitlab.com/gitlab-org/gitlab/-/issues/271575 it was
discovered that the current configuration of how to connect to consul
and obtain the database server list is slightly in-efficent in the amount
of DNS queries that will take place.

With the current configuration, we make 2 dns queries according to the
`/etc/resolv.conf` configuration of the pod.

* consul-consul-dns.consul.gitlab.svc.cluster.local which fails
* consul-consul-dns.consul.svc.cluster.local which then succeeds

This change gives the fully qualified name of the consul Kubernetes
service DNS record we wish to lookup, reducing the amount of dns requests
that need to happen by 1 (50%). While this might not seem like a big change,
anything we can do to reduce the amount of DNS traffic (especially traffic
we know will always fail) helps resource utilisation on `kube-dns` and increases
responsiveness of the application.