Remove `nodeselector` from `kas`
As part of incident https://gitlab.com/gitlab-com/gl-infra/production/-/issues/4099
we want to reduce the number of vms running in `gprd` as much as possible. `kas`
currently only runs in 2 pods (and likely won't grow much for a while), so having
a full node pool just to run it is overkill right now, and we could use shutting down
the vms.