Configure ingress-nginx to keep the real client IP in `X-Forwarded-For`
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1464

In this comment https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1464#note_541153348
it was discovered that when migrating the API service to Kubernetes,
our logs in ES lost the original client IP of the address. This is due
to a nginx configuration difference between VMs and containers.

This change adds some `nginx-ingress` specific settings we need in
order to make sure the real client IP is still preserved in our logs.