Add docker tls volumemount to gitlab-runner jobs
As part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1648
we need the gitlab-runner using the Kubernetes executor to work
building docker images (to test registry functionality).

According to the documentation at

https://docs.gitlab.com/ee/ci/docker/using_docker_build.html#docker-in-docker-with-tls-enabled-in-kubernetes

In order allow this to happen, we need to configure the runner to populate a volumeMount that client
certificates will get populated into for each job. This volume can then be passed to the CI job so that
docker running inside the job can use the client certs to connect to the remote docker daemon using
TLS authentication.