Add a headless consul service temporarily and test using it in `gstg`
Part of https://gitlab.com/gitlab-org/gitlab/-/issues/271575

As documented at https://kubernetes.io/docs/concepts/services-networking/service/#headless-services
it's possible to deply a service with `spec.clusterIP` set to `None`, which
will make the service a headless service. This means no clusterIP is assigned,
and when the dns record for the query is fetched, it will return A records
which are the pod IPs for every pod that matches the service, rather than returning
the ClusterIP itself (as headless services don't have any)

The unfortunate problem is you cannot set a ClusterIP to `None` upon modification.
You can only do it on service creation. This means we can't easily change the current
Consul DNS service to be headless without deleting and re-adding the service, using
two helm invocations, and causing downtime.

Instead what this change does is includes a new headless service for consul in
the `gitlab-extras` helm release, to be available alongside the existing one from the
consul helm chart. This change then configures `gstg` to use this new headless service
for finding consul to query for db replicas.

Assuming there are no issues, we will roll this out into production temporarily.

Changing to the new headless service will bypass using the ClusterIP kube-proxy iptables
rules to route the traffic to consul, instead looking up consul will return multiple A records
which are pod IPs. I assume our ruby library will then just pick a random A record to use and
send the consul dns request to that IP directly. This should allow us to further debug the problems
we are seeing as instead of reporting connectivity issues to the ClusterIP, we will actually see real
pod ip addresses being reported in logs, allowing us to correlate those logs with actual consul pods
in other logs, to further understand what is going on.