Enable `gitlab-extras` release in zonal clusters but conditionally apply values
As part of MR https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-com/-/merge_requests/798
we added a new consul service object to be used by the webservice pods to the
`gitlab-extras` release. Unfortunately this helm release isn't installed on the zonal
clusters, and if it is installed, we do not want the kas CRDs on the zonal clusters
as kas is not running there.