Change the memory requests on gitlab-shell to be the same as the limits.
Part of https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/12767

Due to the way our pods that service Gitlab shell are setup, where openssh
SSHD execs gitlab-shell, we have a very strict memory pattern when we have lots
of connections.

An active SSH connection takes up a large fixed chunk of RAM due to the
architecture of openssh + gitlab-shell, and we have no real controls to limit
the amount of active connections (that is, connections that have made it
through authentication and "doing work").

What this means is during periods of performance degredation or outage to
gitlab-shells dependant services, such as Gitaly or the postgresql database,
these connections keep incoming and because they are slow/timing out, they
stay there taking up memory inside the pod.

This causes rapid memory growth inside the gitlab-shell pod, which eventually
leads to the node being memory exhausted. The kubelet does step in and attempt
to evict the gitlab-shell pod, but the memory growth is so rapid, and the termination
of the gitlab-shell pod is so slow, this causes the Linux OOMKiller to step in,
which starts killing pods randomly, including important system pods like `calico-node`,
which puts the node into such distress it takes a very long time to recover (much longer
than the outage period of the downstream service), and also can cause the GKE
node-problem-detector to step in and fail/rebuild the node, causing more disruption.

To avoid all this, we want to set the requests of the gitlab-shell pod to match
the limit. This will cause the scheduler to schedule gitlab-shell pods in such a way in
that if the pods do all start growing in memory due to a downstream service failure,
the pods will grow to their 1G memory limit and then get terminated, which is bad for the
service, but will prevent the node from running out of memory and OOMKiller stepping in.

It's an unfortunate compromise we need in order to guarentee the health of the node
even during incidents.