Disable specific nginx settings for api ingress-nginx
As part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1683 we audited
the current nginx configuration we used across all our nodes, and determined that
we need specific configuration for certain locations. We need to carry this configuration
over into ingress-nginx.

This change brings that configuration into the ingress nginx configuration in a bit of
a hacky way. But it gets the job done.