Set `.global.appConfig.kas.enabled` to true to fix kas in staging
In https://gitlab.com/gitlab-org/gitlab/-/issues/328946 it was discovered
that kas (which talks to Gitlab API) has been throwing 401 errors since
April 16th. This is due to us switching staging over to have API running
in Kubnernetes. You would think this wouldn't be an issue, but because kas
runs in the regional cluster, and the api pods run in the zonal clusters,
the helm chart configuration for both diverges, so the zonal clusters think
kas is disabled (thus don't sync the secret for kas to authenticate) while
the regional cluster does know kas is enabled.

This change fixes this discrepancy and fixes kas.