Bump chart version to latest
Related to https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1623

Bump the chart version to bring in (currently unused) changes for to the
kas service object, so we can expose it via a LoadBalancer later.

As this change will bring in the workhorse graceful termination
changes, we need to make critically sure we have all our timeout settings
correct.