Add optional tracechunks Redis config if available
Because this is being rolled out, the sentinel config is going to be null until we make some separate chef config updates, which need to be gated on a change issue.

Therefore the diff for this MR will initially be a noop but during https://gitlab.com/gitlab-com/gl-infra/production/-/issues/4884 we will add gstg config, merge this, add gprd config, and then do a CHEF_CONFIG_UPDATE change to kick it over on gprd.  Technically this could actually be merged before, and a CHEF_CONFIG_UPDATE change be done for gstg, but that's extra paperwork