Change websockets and git-https to use nodeSelector based off GKE node pool name
Corrective action for incident gitlab-com/gl-infra/production#4940

As discussed in issue above at gitlab-com/gl-infra/production#4940 (comment 606986184)
Our current setup of setting arbritrary labels on node-pools and using that
as a nodeSelector has limitiations.  If a Pod is subject to 2 node pools, the cluster auto-scaler fails
to scale either pool up, which leads to a situation where unschedulable Pods start to appear, which fails a deploy

This change modifys the websockets and git-https deployments to use nodePool names instead of our arbritrary labels,
making sure to use the correct node-pool names in each environment.