Modify `gitlab-extras` so `ingress/gitlab-readiness` is not installed where it isn't needed
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1809

As discovered during investigation of incident https://gitlab.com/gitlab-com/gl-infra/production/-/issues/4962
we actually install this ingress into our regional clusters for staging and production, where it actually
isn't needed. On top of this, in production, we install what is essentially a duplicate ingress object
into both the canary and non-canary namespaces, which might be confusing our metrics (it could not be,
but cleaning this up is one way to make sure it isn't).