Remove `type: api` label from gitlab-readiness ingress
Corrective action for https://gitlab.com/gitlab-com/gl-infra/production/-/issues/4962

It seems that the reason our metrics for api disappeared is that our
monitoring setup was assuming that we had a singular ingress object
for `type:api`. Before we added `ingress-readiness` we did, and this
made sense.

As this ingress object really isn't just for api (it will be also for
web eventually), we should remove this label for the moment to fix
our metrics.