Route catchall sidekiq jobs to default queue in gstg
In pursuit of https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/447 we want to do all of gstg catchall in one go.

The routing has been tested, this is just a mechanical exercise, in preparation for production which will be done
in batches (https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1074) for risk-mitigation.