[prod] Change kas to use `frontendConfig` object to make GKE ingress for kas do HTTP to HTTP(S) redirects
Closes: https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1755

Corrective action for: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5108

Unfortunately the annotation `allow.http` on ingress objects breaks Google
Managed Certificates and shouldn't be used with GKE ingress and Google Managed
Certificates (which is what we use for KAS).

Google instead recommend we create a `frontendConfig` object specifying HTTP
to HTTP(S) redirect and apply that to the ingress in question instead.

This change is for production, and has been tested to not cause an outage when
applied https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1755#note_623400649