Bump all webservice pods in pre to minReplicas of 2
Corrective action for https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5125

We currently use preemptive instances in pre, and the GKE notes specify that when
an instance is preempted, it violates all Kubernetes safeguards like PDBs and shutdown
grace periods. This means if we are running a single pod and it gets preempted, we will
have downtime on that service.

This likely has been happening for a while but we never really noticed until we migrated
web in pre to Kubernetes, as the blackbox probes will start to fail sometimes because of
this.