Increase minreplicas for gitlab-shell, websockets, and ingress-nginx
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1874

As described in the GKE Notes on pre-emptible instances

https://cloud.google.com/kubernetes-engine/docs/how-to/preemptible-vms

When an instance is preempted, it doesn't follow the normal Kubernetes
safeguards to ensure things are shutdown cleanly and traffic is not
disrupted.

To this end, I have identified the following deployments that run only
1 pod of their type in pre and thus are susceptable to outages when instance
preemption happens.

This change bumps them to have at least two pods minimum.

On top of this, as ingress-nginx has no inbuilt antiAffinity, we add this in
pre so that both pods aren't scheduled onto the same node (and then preempted).