Convert service for web component from type `ClusterIP` to type `LoadBalancer`
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1924

This is the first step in removing nginx from being in front of web. We will
convert the Kubernetes service to be exposed as type `LoadBalancer` so that
haproxy can directly connect to it.

This change is for production, following on from

https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-com/-/merge_requests/1078