Disable ingress-nginx from regional cluster in gprd
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1916

We have identified that ingress-nginx is not actually needed in our
regional cluster in gstg and gprd (but needed for gprd-cny).

In order to simplify things and make sure ingress controllers aren't
picking up objects they shouldn't, we will disable this in gprd only.