Change networkpolicy for web pods to allow API nodes to talk to KAS
Part of https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/13821

As part of upcoming features for the Gitlab product

* https://gitlab.com/gitlab-org/gitlab/-/issues/322605

The API pods need to talk to Kas pods. We have already setup an internal
load balancer at `int.${environment}.gitlab.net` on port 8153 which
forwards to the Kas internal endpoints.

We have configured the API nodes to know of this address/port/protocol,
we just need to allow the pods connectivity at the network level.

This change is for nonprod only.