Add a dummy queue to the end of the catchall list
For incident https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5357, adds a 'sacrificial' queue that will get a spurious single quote added to it, to allow the other active queues to be listened for correctly.

This is temporary only, until the underlying bug (precise details unknown) can be carefully corrected with sufficient review.