Increase number of workers, and resources given to webservice pod in grpd for us-east1-b only
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1873

Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1894

After the initial migration and stabilisation of web running on Kubernetes
in production, we have noticed that we running far more pods than we would
like, and our performance metrics for request `duration_s` and `queue_duration_s`
is higher than what we previously had in VMs.

Going off what we have learnt from tuning api nodes in Kubernetes, we want to try
increasing the number of workerProcesses in web nodes from 4 to 6, and increasing
the number resources requests/limits to allow for the new threads to function
without starvation.

This is for gprd-us-east1-b only, for us to distinctly monitor it is having a positive
effect, before rolling out to other clusters.