Disables workhorse.keywatcher on all envs
Target is gprd; as this was previously done explicitly to gstg (removed in this commit) gstg will be a noop, but it'll change gprd + pre.  This is a performance fix for Redis; https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/12812 describes the original intent to only listen to these keys on API, and https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1261 the investigation into CPU saturation stress on Redis after migrating web to k8s, which this should resolve.