Vendor `raw` chart into repo
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1982

In order to allow us to eventually drop `helm-git`, as well as remove
external depedencies and help readability, we are moving external charts
to be vendored in repo.

As the raw chart is currently being hosted on the incubator repository
that is locked (no more changes), we can easily move this in repo completely
(no need to track it upstream using `git subtree`, as it will never change).