Add urgent_other queue and explicitly list urgent-other queues
We are adding a new `urgent_other` queue which will be used for all
jobs (except those in the `email_receiver` queue) on the `urgent-other`
shard. As the first step of this change, we need to explicitly list this
new `urgent_other` queue alongside all the existing queues, to avoid
losing jobs as we re-route them.

This is the production equivalent of
https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-com/-/merge_requests/1159