Add quarantine, gitaly_throttled, and database_throttled queue and list implied queues
To migrate the 3 shards to one-queue-per-shard, we are adding a new queue for
each shard which will be used for all jobs on that shard. As the first step
of this change, we need to explicitly name the new queue alongside all the
existing queues, to avoid losing jobs as we re-route them.