Remove redundant sidekiq routing rules
While migrating the catchall shard to one-queue-per-shard, we created interim rules that incrementally routed work to the default shard.  With the completion of that and replacing the final rule with a wildcard routing to default, the transient rules are no longer required.  This is a noop on where jobs actually run.