Move service_desk_email_receiver to urgent_other
For https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1207, moves this worker to a more logical shard.  It was previously on catchall more by co-incidence than intent, and it still needs its own queue because of the source (mailroom).  See https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1087 for a brief discussion of that.

Moves in one single MR for both envs because:
1. Actual rate of jobs is incredibly low (in gprd maybe a couple of hundred a day, clustered on EU/AMER hours, with typically several hours at a time during quiet periods where it receives no use
1. Short (10-15 minute) delays (worst case, if catchall stopped listening before  urgent-other started) is quite acceptable; we've had previous incidents ( https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5212) where jobs were queued for days and it wasn't ideal but wasn't a disaster.  Plus it feeds of SMTP which is already notoriously non-real-time.  Doing 4 MRs (2 per env to expand/contract) would be inefficient for no material gain.