Change Gitlab CI detection of a chart bump taking place
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2031

As part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1982
we changed our charts used in this repo to be vendored inside the repo.

However, I missed the fact that we need to change our Gitlab CI configuration
on what files to match when determining if we need to run QA jobs or not.

This change corrects this.