Add `charts/gitlab/**` to changes stanza for QA jobs
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2031

While this was added to the .trigger-qa-smoke job in MR

https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-com/-/merge_requests/1262

it seems this didn't have the desired effect. Likely because the
actual QA jobs overwrite the `changes:` stanza.

This change adds the `charts/gitlab/**` stanza directly to the QA jobs
themselves.