Bump CI image to include full version of ps
Part of gitlab-com/gl-infra/delivery#1626

With MR

https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-com/-/merge_requests/1278

we introduced a watch-gitlab-deployments.sh script to give us some output
about our helm deployments when they are happening. However this script relies
on a full version of ps to function, not the stripped down version in busybox
with the image currently uses.

This change installs the full version so the script functions correctly.