Move all gcloud addresses and parts of chef configuration to be directly captured.
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1960

With the current status of the Gitlab.com Kubernetes migration, along with
the recent issues we have with unwanted changes being introduced by external
sources having issues, we want to simplify things by making the settings we
used to sync from chef be instead directly configured in this repository instead.

This is not a complete migration, but is a first pass and attempts to capture
the easiest settings to migrate from syncing to directly defined.