Add `type` label to all gitlab-pages objects
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1969

After looking at the pages metrics we can see that some are still empty

This is because the metrics expect a `type: web-pages` label on all metrics.

This is missing because we don't have that label on our pages Kubernetes objects.