Temporarily raise memory-bound replicas
To help debug https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5683 we want to force a 4th node in the memory-bound nodepool to see if the problem exists there.  The nodes are n1-standard-8 (30GB RAM) and each pod has a 3GB memory request, so we need somewhere approaching 30 to force a 4th node to be spun up.  32 is a nice round number sufficiently in excess of the minimum to be 100% confident it will have the desired effect.   The node pool has a limit of 10 nodes, so will not be a constraint.