feat(workhorse): enable graceful shutdown in gstg
What
---
Set `GITLAB_WORKHORSE_EXEC` that was introduced in
https://gitlab.com/gitlab-org/build/CNG/-/merge_requests/972 so that
`gitlab-workhorse` is PID `1` and `SIGTERM` is sent.

Why
---
In
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15497#note_910106152
we see that the `gitlab-workhorse` container takes 30 seconds to
shutdown, which is the full 30 seconds of `terminationGracePeriod` which
is by default 30 seconds. In
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15497#note_911022441
we found out that `gitlab-workhorse` process is not triggering a
graceful shutdown and we end up sending `SIGKILL` to the process.

In those 30s we can still end up serving requests because it's not yet
removed from the service, and we end up killing requests that are
in-flight. Improving the container shutdown might improve the chance of
us not serving any requests when puma is already shutdown.

Testing
---
- [Termination logs](https://nonprod-log.gitlab.net/goto/6520ee20-bf18-11ec-b3a6-472d0398dd6e)
- [workhorse 502 errors](https://nonprod-log.gitlab.net/goto/a3f7a940-bf18-11ec-b3a6-472d0398dd6e)
- [pages 502 errors](https://nonprod-log.gitlab.net/goto/be346d70-bf18-11ec-b3a6-472d0398dd6e)

Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>