fix(workhorse): graceful shutdown gprd-us-east1-b
What
---
Set `GITLAB_WORKHORSE_EXEC` that was introduced in
https://gitlab.com/gitlab-org/build/CNG/-/merge_requests/972 so that
`gitlab-workhorse` is PID `1` and `SIGTERM` is sent to the workhorse
process which will trigger the graceful shutdown.

Why
---
In
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15497#note_910106152
we see that the `gitlab-workhorse` container takes 30 seconds to
shutdown, which is the full 30 seconds of `terminationGracePeriod` which
is by default 30 seconds. In
https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15497#note_911022441
we found out that `gitlab-workhorse` process is not triggering a
graceful shutdown and we end up sending `SIGKILL` to the process.

In those 30s we can still end up serving requests because it's not yet
removed from the service, and we end up killing requests that are
in-flight. Improving the container shutdown might improve the chance of
us not serving any requests when puma is already shutdown, and
gracefully close http connections.

Testing
---

We've rolled this out in
[`gstg`](https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-com/-/merge_requests/1714#note_917747951)
and
[`gprd-cny`](https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-com/-/merge_requests/1720#note_918357983)
it shows good results which eliminates the `502` errors.

Reference: https://gitlab.com/gitlab-com/gl-infra/reliability/-/issues/15497
Signed-off-by: Steve Azzopardi <sazzopardi@gitlab.com>