Enable correlation ID propagation for Git over SSH requests
The Webservice containers used to serve gitlab-shell had the Workhorse
`trusted_cidrs_for_propagation` setting, but it did not have the
required `-propagateCorrelationID` flag in Workhorse to enable this
feature.

As a result, gitlab-shell would generate a correlation ID, but
Workhorse would generate a new one because this flag was not enabled.
We already use this feature in production for Web requests; SSH
requests would benefit from this as well.