Change rack-attack to be pulled from Kubernetes secrets in prod
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2371

This change introduces an new secret which is the IP addreses that
are whitelisted in rack attack (so not to be rate limited). This
is currently passed as a helm value directly into the chart, but
with this change, we will put the same data into a Kubernetes
secret instead, then modify the pods in such a way that the
configure init container will pull the contents of the secret
(which is mounted on disk in the pod) into the final gitlab.yml
file to be used by the container.

MR for non-prod at https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-com/-/merge_requests/1816