Move to using `vendir` for vendoring charts
Now that the upstream charts actuall vendor their upstream
dependencies directly (due to bitnami repo issues), we can
greatly clean and simplify the process around vendoring the
upstream chart into this repo.

To do this, we switch to using the [vendir](https://carvel.dev/vendir/)
which gives us a much simpler tool and workflow for vendoring
the chart and filtering out resources, drift detection, etc.

Part of this we can switch to using the `vendor` directory
for vendored pieces, to make this more obvious.