Correct CI configuration for new vendor setup
In MR https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-com/-/merge_requests/1871
we changed where and how the charts are vendored into this repository.

However as part of that change we needed to also adjust the paths of
where the charts are located as part of `changes` stanza. This was
missed and thus this change fixes it.