Disable Puma worker killer everywhere
It has not been running for a while due to a bug
in how the Rails app interpreted this switch:
gitlab-org/gitlab!88928

Since the value of this is dubious in face of existing
container memory limits, and to de-risk deploying the
fix without knowing what the new limits should be,
I think we should disable this explicitly for now.

This change disables it in all environments.