Go back to manually vendoring subcharts
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/2459

With MR https://gitlab.com/gitlab-org/charts/gitlab/-/merge_requests/2607/
in the upstream charts, we can no longer rely on the fact that the
subcharts for the GitLab chart are vendored into the Git repo.

This means we unfortunately need to back to vendoring subcharts
manually, using `helm dep update`.

This change will update the documentation (for manual chart bumps)
and the CI jobs to accomodate this.